+++ include/Prokaryote.hh
@@ -34,7 +34,7 @@ class Prokaryote{
 
 		void InitialiseProkaryote();
 		void ClonePPFromPP(Prokaryote* PP_template, int tot_prok_count);
-		void Replicate(int env, int res);
+		void Replicate(double env, int res);
 		void Mitosis(Prokaryote* parent, int tot_prok_count);
 		void Abortion();
 		void EmptyProkaryote();
